#56 - platform into libmu namespace
