+++ src/libmu/platform/platform.cc
@@ -1,35 +1,10 @@
-/*******
+/********
  **
- ** Copyright (c) 2017, James M. Putnam
- ** All rights reserved.
+ **  SPDX-License-Identifier: MIT
  **
- ** Redistribution and use in source and binary forms, with or without
- ** modification, are permitted provided that the following conditions are met:
+ **  Copyright (c) 2017-2020 James M. Putnam <putnamjm.design@gmail.com>
  **
- ** 1. Redistributions of source code must retain the above copyright notice,
- **    this list of conditions and the following disclaimer.
- **
- ** 2. Redistributions in binary form must reproduce the above copyright
- **    notice, this list of conditions and the following disclaimer in the
- **    documentation and/or other materials provided with the distribution.
- **
- ** 3. Neither the name of the copyright holder nor the names of its
- **    contributors may be used to endorse or promote products derived from
- **    this software without specific prior written permission.
- **
- ** THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- ** AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- ** IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- ** ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
- ** LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- ** CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- ** SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- ** INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- ** CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- ** ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- ** POSSIBILITY OF SUCH DAMAGE.
- **
- *******/
+ **/
 
 /********
  **
@@ -49,11 +24,12 @@
 #include <time.h>
 #include <unistd.h>
 
-#include "platform/ffi.h"
-#include "platform/platform.h"
+#include "libmu/platform/platform-ffi.h"
+#include "libmu/platform/platform.h"
 
 extern char **environ;
 
+namespace libmu {
 namespace platform {
 
 /** * map pages in system address space **/
@@ -117,5 +93,5 @@ std::string Platform::Invoke(uint64_t fnp, std::string arg) {
 
   return fn(arg);
 }
-
 } /* namespace platform */
+} /* namespace libmu */
